requirements.txt
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
diffusers
gradio
fastapi
uvicorn
requests
pillow
